+++ CHANGELOG.md
@@ -1,3 +1,2 @@
 ## Changelog
-- sync upstream `26.2-0.29.6`
-- update NeoForge version `26.2.0.37-beta` -> `26.2.0.79`
+- rework network part, use NeoForge native API